# src/test/resources/openapi/approved-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "technical-interview-demo API",
    "description": "Machine-readable contract for the demo application's supported external /api/** surface. Internal-only overview, documentation, OpenAPI publication, and actuator validation paths are intentionally excluded.",
    "version": "APPROVED"
  },
...
  "paths": {
    "/api/localizations/{id}": {
      "get": {
        "tags": [
          "Localizations"
        ],
        "summary": "Get a localization by id",
        "description": "Public endpoint for a single localization.",
        "operationId": "findById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LocalizationResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Localizations"
        ],
        "summary": "Update a localization",
        "description": "Requires an authenticated session with the ADMIN role and a valid same-site CSRF header mirrored from the readable XSRF-TOKEN cookie.",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "X-XSRF-TOKEN",
            "in": "header",
            "description": "Same-site CSRF header whose value must match the readable XSRF-TOKEN cookie.",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LocalizationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Localizations"
        ],
        "summary": "Delete a localization",
        "description": "Requires an authenticated session with the ADMIN role and a valid same-site CSRF header mirrored from the readable XSRF-TOKEN cookie.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "X-XSRF-TOKEN",
            "in": "header",
            "description": "Same-site CSRF header whose value must match the readable XSRF-TOKEN cookie.",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/books/{id}": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get a book by id",
        "description": "Public endpoint that returns a single book with its assigned categories.",
        "operationId": "findById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Books"
        ],
        "summary": "Update a book",
        "description": "Requires an authenticated session, a valid same-site CSRF header mirrored from the readable XSRF-TOKEN cookie, and the current optimistic-lock version.",
        "operationId": "update_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "X-XSRF-TOKEN",
            "in": "header",
            "description": "Same-site CSRF header whose value must match the readable XSRF-TOKEN cookie.",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Books"
        ],
        "summary": "Delete a book",
        "description": "Requires an authenticated session and a valid same-site CSRF header mirrored from the readable XSRF-TOKEN cookie.",
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "X-XSRF-TOKEN",
            "in": "header",
            "description": "Same-site CSRF header whose value must match the readable XSRF-TOKEN cookie.",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/books": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "List books",
        "description": "Public endpoint with pagination, sorting, text filters, category filters, and publication-year filters.",
        "operationId": "findAll_2",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "description": "Case-insensitive substring filter for the title.",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Case-insensitive substring filter for the title.",
              "example": "clean"
            },
            "example": "clean"
          },
          {
            "name": "author",
            "in": "query",
            "description": "Case-insensitive substring filter for the author.",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Case-insensitive substring filter for the author.",
              "example": "bloch"
            },
            "example": "bloch"
          },
          {
            "name": "isbn",
            "in": "query",
            "description": "Case-insensitive substring filter for the ISBN.",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Case-insensitive substring filter for the ISBN.",
              "example": "978013"
            },
            "example": 978013
          },
          {
            "name": "year",
            "in": "query",
            "description": "Exact publication year. Cannot be combined with yearFrom or yearTo.",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Exact publication year. Cannot be combined with yearFrom or yearTo.",
              "format": "int32",
              "example": 2018
            },
            "example": 2018
          },
          {
            "name": "yearFrom",
            "in": "query",
            "description": "Inclusive publication-year lower bound.",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Inclusive publication-year lower bound.",
              "format": "int32",
              "example": 2000
            },
            "example": 2000
          },
          {
            "name": "yearTo",
            "in": "query",
            "description": "Inclusive publication-year upper bound.",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Inclusive publication-year upper bound.",
              "format": "int32",
              "example": 2020
            },
            "example": 2020
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "description": "Repeat the parameter to match any of multiple categories.",
              "items": {
                "type": "string",
                "description": "Category filter value.",
                "example": "Java"
              }
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [
                "id,ASC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageBook"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Books"
        ],
        "summary": "Create a book",
        "description": "Requires an authenticated session established through the configured OAuth provider login flow and a valid same-site CSRF header mirrored from the readable XSRF-TOKEN cookie.",
        "operationId": "create_2",
        "parameters": [
          {
            "name": "X-XSRF-TOKEN",
            "in": "header",
            "description": "Same-site CSRF header whose value must match the readable XSRF-TOKEN cookie.",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "LocalizationRequest": {
        "required": [
          "language",
          "messageKey",
          "messageText"
        ],
        "type": "object",
        "properties": {
          "messageKey": {
            "maxLength": 150,
            "minLength": 0,
            "pattern": "^[a-z0-9._-]+$",
            "type": "string",
            "description": "Stable message key managed by the application.",
            "example": "error.request.resource_not_found"
          },
          "language": {
            "minLength": 1,
            "pattern": "^[a-zA-Z]{2}$",
            "type": "string",
            "description": "Two-letter ISO 639-1 language code.",
            "example": "pl"
          },
          "messageText": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string",
            "description": "Localized message text.",
            "example": "Nie znaleziono zasobu."
          },
          "description": {
            "maxLength": 1000,
            "minLength": 0,
            "type": "string",
            "description": "Optional reviewer-facing description.",
            "example": "Displayed when a requested resource does not exist."
          }
        },
        "description": "Payload for creating or updating a localization message."
      },
      "LocalizationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Database identifier.",
            "format": "int64",
            "example": 1
          },
          "messageKey": {
            "type": "string",
            "description": "Stable message key.",
            "example": "error.request.resource_not_found"
          },
          "language": {
            "type": "string",
            "description": "Two-letter ISO 639-1 language code.",
            "example": "en"
          },
          "messageText": {
            "type": "string",
            "description": "Localized message text.",
            "example": "The requested resource was not found."
          },
          "description": {
            "type": "string",
            "description": "Optional reviewer-facing description."
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp as a UTC instant.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last update timestamp as a UTC instant.",
            "format": "date-time"
          }
        },
        "description": "Localization message returned by the API."
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Database identifier.",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Unique category name.",
            "example": "Java"
          }
        },
        "description": "Book category."
      },
...
      "BookUpdateRequest": {
        "required": [
          "author",
          "publicationYear",
          "title",
          "version"
        ],
        "type": "object",
        "properties": {
          "title": {
            "minLength": 1,
            "type": "string",
            "description": "Book title.",
            "example": "Effective Java, Third Edition"
          },
          "author": {
            "minLength": 1,
            "type": "string",
            "description": "Primary author.",
            "example": "Joshua Bloch"
          },
          "version": {
            "type": "integer",
            "description": "Current optimistic-lock version.",
            "format": "int64",
            "example": 0
          },
          "publicationYear": {
            "type": "integer",
            "description": "Publication year.",
            "format": "int32",
            "example": 2018
          },
          "categories": {
            "type": "array",
            "description": "Optional replacement list of existing category names.",
            "items": {
              "type": "string",
              "description": "Existing category name.",
              "example": "Java"
            }
          }
        },
        "description": "Payload for updating an existing book. ISBN remains immutable."
      },
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Database identifier.",
            "format": "int64",
            "example": 1
          },
          "version": {
            "type": "integer",
            "description": "Optimistic-lock version used for updates.",
            "format": "int64",
            "example": 0
          },
          "title": {
            "type": "string",
            "description": "Book title.",
            "example": "Effective Java"
          },
          "author": {
            "type": "string",
            "description": "Primary author.",
            "example": "Joshua Bloch"
          },
          "isbn": {
            "type": "string",
            "description": "Unique ISBN assigned at creation time.",
            "example": "9780134685991"
          },
          "publicationYear": {
            "type": "integer",
            "description": "Publication year.",
            "format": "int32",
            "example": 2018
          },
          "categories": {
            "type": "array",
            "description": "Assigned categories ordered by name.",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          }
        },
        "description": "Book resource returned by the API."
      },
...
      "BookCreateRequest": {
        "required": [
          "author",
          "isbn",
          "publicationYear",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "minLength": 1,
            "type": "string",
            "description": "Book title.",
            "example": "Effective Java"
          },
          "author": {
            "minLength": 1,
            "type": "string",
            "description": "Primary author.",
            "example": "Joshua Bloch"
          },
          "isbn": {
            "minLength": 1,
            "type": "string",
            "description": "Unique ISBN for the new book.",
            "example": "9780134685991"
          },
          "publicationYear": {
            "type": "integer",
            "description": "Publication year.",
            "format": "int32",
            "example": 2018
          },
          "categories": {
            "type": "array",
            "description": "Optional list of existing category names.",
            "items": {
              "type": "string",
              "description": "Existing category name.",
              "example": "Java"
            }
          }
        },
        "description": "Payload for creating a new book."
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "paged": {
            "type": "boolean"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
      "PageBook": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Book"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "description": "Authenticated browser session cookie used by protected operations. It is established through a configured identity provider login path under /api/session/oauth2/authorization/{registrationId} when the optional oauth profile is active. Unsafe browser writes also require the X-XSRF-TOKEN request header mirrored from the readable XSRF-TOKEN cookie.",
        "name": "technical-interview-demo-session",
        "in": "cookie"
      }
    }
  }
}